Feat/88 create shared food

+++ src/main/java/no/ntnu/stud/idatt2106/backend/model/request/FoodRequest.java
@@ -12,7 +12,7 @@ public class FoodRequest {
   private int typeId;
   private int householdId;
   private LocalDate expirationDate;
-  private int amount;
+  private float amount;
 
   public int getTypeId() {
     return typeId;
@@ -38,11 +38,11 @@ public class FoodRequest {
     this.expirationDate = expirationDate;
   }
 
-  public int getAmount() {
+  public float getAmount() {
     return amount;
   }
 
-  public void setAmount(int amount) {
+  public void setAmount(float amount) {
     this.amount = amount;
   }
 }
